Add files via upload
The project used a methodical approach, which included pre-processing of the data, descriptive statistics, and the use of K-means grouping. The analysis's findings and results have given us important information for making focused marketing strategies.
At the data pre-processing stage, missing values were handled, the data was cleaned up, and summary statistics were used. This step made sure that the data set was clean and easy to understand, so that it could be analysed in a useful way. K-means clustering was picked as the clustering method because it works well to find groups of data that are different from each other.
The clustering research led to useful results, such as figuring out different groups or segments of customers. Each cluster had its own characteristics, interests, and ways of acting. When companies know about these groups, they can tailor their marketing to meet the needs and wants of each segment. This makes it possible to make marketing efforts that are more relevant to customers. This improves engagement, customer satisfaction, and, in the end, profits.
The ideas and patterns that came out of the clustering analysis are useful for making decisions about marketing strategies. When companies know how to divide their customers into groups, they can tailor their marketing strategies, make relevant product suggestions, and give customers a more personalised experience. By matching marketing efforts to the traits of each cluster, businesses can make the best use of their resources, improve their relationships with customers, and earn their long-term trust.
In the end, this mini-project on data mining has shown how important it is to use data-driven methods to get insights from large datasets. Using K-means clustering, and other pre-processing methods has made it possible to analyse the e-commerce dataset in depth. The results and suggestions from this analysis can be used as a basis for targeted marketing strategies that help businesses reach and connect their customers more effectively, which leads to better business outcomes.